added opcode interpreter

+++ src/interpreter.cpp
@@ -2,102 +2,100 @@
 
 #include <stdexcept>
 
-#include "instructions.h"
-
 namespace tci {
 void Interpreter::Load(Instruction ins, size_t &icounter) {
   auto vname = std::get<std::string>(ins.i0);
-  auto ridx = std::get<size_t>(ins.i1);
+  auto ridx = std::get<int>(ins.i1);
   m_registers[ridx] = m_st.GetValue(vname);
   ++icounter;
 }
 
 void Interpreter::Loadi(Instruction ins, size_t &icounter) {
   auto val = std::get<double>(ins.i0);
-  auto ridx = std::get<size_t>(ins.i1);
+  auto ridx = std::get<int>(ins.i1);
   m_registers[ridx] = val;
   ++icounter;
 }
 
 void Interpreter::Store(Instruction ins, size_t &icounter) {
-  auto ridx = std::get<size_t>(ins.i0);
+  auto ridx = std::get<int>(ins.i0);
   auto vname = std::get<std::string>(ins.i1);
   m_st.SetValue(vname, m_registers[ridx]);
   ++icounter;
 }
 
 void Interpreter::Add(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] + m_registers[s1];
   ++icounter;
 }
 
 void Interpreter::Sub(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] - m_registers[s1];
   ++icounter;
 }
 
 void Interpreter::Mul(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] * m_registers[s1];
   ++icounter;
 }
 
 void Interpreter::Div(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] / m_registers[s1];
   ++icounter;
 }
 
 void Interpreter::Gt(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] > m_registers[s1] ? 1.0 : 0.0;
   ++icounter;
 }
 
 void Interpreter::Lt(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] < m_registers[s1] ? 1.0 : 0.0;
   ++icounter;
 }
 
 void Interpreter::Lteq(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] <= m_registers[s1] ? 1.0 : 0.0;
   ++icounter;
 }
 
 void Interpreter::Gteq(Instruction ins, size_t &icounter) {
-  auto s0 = std::get<size_t>(ins.i0);
-  auto s1 = std::get<size_t>(ins.i1);
-  auto d = std::get<size_t>(ins.i2);
+  auto s0 = std::get<int>(ins.i0);
+  auto s1 = std::get<int>(ins.i1);
+  auto d = std::get<int>(ins.i2);
   m_registers[d] = m_registers[s0] >= m_registers[s1] ? 1.0 : 0.0;
   ++icounter;
 }
 
 void Interpreter::Jmp(Instruction ins, size_t &icounter) {
-  auto offset = std::get<size_t>(ins.i0);
+  auto offset = std::get<int>(ins.i0);
   icounter = icounter + offset;
 }
 
 void Interpreter::CJmp(Instruction ins, size_t &icounter) {
-  auto ridx = std::get<size_t>(ins.i0);
-  auto offset = std::get<size_t>(ins.i1);
+  auto ridx = std::get<int>(ins.i0);
+  auto offset = std::get<int>(ins.i1);
   if (m_registers[ridx] == 1.0) {
     icounter = icounter + offset;
   } else {
